foji_project/requirements.txt
Django
djangorestframework
celery

openpyxl
redis

django-smtp-ssl

gunicorn

factory_boy
faker
django_extensions
